Rename 'final' module to 'finalized' to avoid name collision with 'final' decorator
